backend/requirements.txt
e2b_code_interpreter
fastapi
uvicorn
openai 
python-dotenv